LazyEncodedString everywhere (#247)
* Massive update to support LazyEncodedString Everywhere
* Not complete!
* This compiles the interop and wrapper libraries only, no impacted libs are updated yet.

* Completed changes with updates to consumers
* Includes additional testing

* More changes to support LazyEncodedString everywhere.

* Updated doc comments on `DisposeMEssageMarshaller type for clarity

* Attempting to work around problems in JIT on Windows x64.
* Apparently the JIT is assuming a 32 bit relocation as the default.
    - See https://github.com/llvm/llvm-project/issues/65641
* This adds support for JIT builder and, in particular, TargetMachineBuilder so that the CodeModel is explicit in children
* Static usings are now explicit everywhere they are needed.
    - This removes static usings from the global usings.

* Additional support for LazyEncodedString
* Converted name of static class for LLVM-C interop `Error` for consistency
* Added marshaling form of `LLVMCreateStringError()`
* Added `LLVMErrorRef.CreateForNativeOut()`
    - Eliminates need for annoying IDisposable warning suppression in many cases.
    - This support out/return of native handle for reverse P/Invoke scenarios.
* Added setting the triple to the module transform for Kaleidoscope
    - Ensures the triple is set and matches the JIT.
* Added `MarshalGCHandle.TryGet<>()` to support more intuitive acquisition of a managed context in native callbacks.
* Moved static factory for `ErrorInfo` to a constructor.
   - Static factory was used for out/return to native but that causes a lot of suppression of warnings for dispose so isn't used. Instead, most APIs will use the static methods on LLVMErrorRef to report any errors. If the callback includes calling through a delegate, then it might use the constructor to build the resulting `ErrorInfo` to pass to native code.

* Updated version checks in interop code
* Since native API surface changed to support this build the version bump is needed

* Additional support for LazyEncodedString
* Added LibLLVMAttributeAllowedOn_All to make testing easier
* Changed signature of LibLLVMGetAttributeInfo to use an out param as the result instead of a raw pointer
    - Wrapper layout is identical to the interop form so a simple cast through a pointer is all that is needed to "project"
* Doc comment updates/fixes

* [Build Fix] Corrected test
* Test file was removed locally and missed
    - commit only included staged files so this edit and replace got in the way.

---------

+++ src/Ubiquity.NET.Llvm/Blake3Hash.cs
@@ -33,7 +33,7 @@ namespace Ubiquity.NET.Llvm
             {
                 fixed(byte* p = &MemoryMarshal.GetArrayDataReference(array))
                 {
-                    llvm_blake3_hasher_update(ref NativeHasher, p + ibStart, cbSize);
+                    llvm_blake3_hasher_update(ref NativeHasher, p + ibStart, (nuint)cbSize);
                 }
             }
         }
